gcc: Support the new riscv{32,64}-{linux,elf} RISC-V tuples
This causes riscv64 to use the riscv directory, which means that the
other RISC-V variants will probably end up using that as well.  I
think this is probably the right thing to do: it's what Intel does,
but not what ARM or Tilera do, but I think that's probably OK because
riscv* should be fairly similar (as opposed to aarch64, which IIRC is
pretty different).